pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-celery-beat = "^2.4.0"
djangorestframework-api-key = ">=2.0.0,<3.0.0"
django-cors-headers = "^3.13.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.10.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'

[tool.isort]
profile = "black"